# live-auction-bid-backend/api/auction/service/v1/user_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v4.25.8
// source: api/auction/service/v1/user.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationUserServiceAdminCreateUser = "/auction.service.v1.UserService/AdminCreateUser"
const OperationUserServiceAdminResetUserPassword = "/auction.service.v1.UserService/AdminResetUserPassword"
const OperationUserServiceAdminUpdateUserRole = "/auction.service.v1.UserService/AdminUpdateUserRole"
const OperationUserServiceAdminUpdateUserStatus = "/auction.service.v1.UserService/AdminUpdateUserStatus"
const OperationUserServiceGetMe = "/auction.service.v1.UserService/GetMe"
const OperationUserServiceListUsers = "/auction.service.v1.UserService/ListUsers"
const OperationUserServiceLogin = "/auction.service.v1.UserService/Login"
const OperationUserServiceLogout = "/auction.service.v1.UserService/Logout"
const OperationUserServiceRefreshToken = "/auction.service.v1.UserService/RefreshToken"
const OperationUserServiceRegister = "/auction.service.v1.UserService/Register"
const OperationUserServiceRegisterMerchant = "/auction.service.v1.UserService/RegisterMerchant"
const OperationUserServiceResetPassword = "/auction.service.v1.UserService/ResetPassword"

type UserServiceHTTPServer interface {
	// AdminCreateUser 主账号创建团队子账号；不创建买家或新的主账号。
	AdminCreateUser(context.Context, *AdminCreateUserRequest) (*AdminCreateUserReply, error)
	// AdminResetUserPassword 主账号重置团队子账号密码；不允许重置买家、主账号或其他商家空间账号。
	AdminResetUserPassword(context.Context, *AdminResetUserPasswordRequest) (*AdminResetUserPasswordReply, error)
	// AdminUpdateUserRole 主账号修改团队子账号角色；不允许改为买家或新的主账号。
	AdminUpdateUserRole(context.Context, *AdminUpdateUserRoleRequest) (*AdminUpdateUserRoleReply, error)
	// AdminUpdateUserStatus 主账号启用或禁用团队子账号。
	AdminUpdateUserStatus(context.Context, *AdminUpdateUserStatusRequest) (*AdminUpdateUserStatusReply, error)
	// GetMe 查看当前登录用户。
	GetMe(context.Context, *GetMeRequest) (*GetMeReply, error)
	// ListUsers 主账号查询团队子账号列表。
	ListUsers(context.Context, *ListUsersRequest) (*ListUsersReply, error)
	// Login 用户名密码登录。
	Login(context.Context, *LoginRequest) (*LoginReply, error)
	// Logout 登出当前 refresh session。
	Logout(context.Context, *LogoutRequest) (*LogoutReply, error)
	// RefreshToken 刷新 access token，并轮换 refresh token。
	RefreshToken(context.Context, *RefreshTokenRequest) (*RefreshTokenReply, error)
	// Register 公开注册买家账号。
	Register(context.Context, *RegisterRequest) (*RegisterReply, error)
	// RegisterMerchant 公开注册商家/主播主账号。
	RegisterMerchant(context.Context, *RegisterMerchantRequest) (*RegisterMerchantReply, error)
	// ResetPassword 原型阶段：仅凭账号重置密码，方便测试和压测准备账号。
	ResetPassword(context.Context, *ResetPasswordRequest) (*ResetPasswordReply, error)
}

func RegisterUserServiceHTTPServer(s *http.Server, srv UserServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/api/users/register", _UserService_Register0_HTTP_Handler(srv))
	r.POST("/api/users/login", _UserService_Login0_HTTP_Handler(srv))
	r.POST("/api/users/reset-password", _UserService_ResetPassword0_HTTP_Handler(srv))
	r.POST("/api/merchants/register", _UserService_RegisterMerchant0_HTTP_Handler(srv))
	r.POST("/api/users/refresh", _UserService_RefreshToken0_HTTP_Handler(srv))
	r.POST("/api/users/logout", _UserService_Logout0_HTTP_Handler(srv))
	r.GET("/api/users/me", _UserService_GetMe0_HTTP_Handler(srv))
	r.POST("/api/admin/users", _UserService_AdminCreateUser0_HTTP_Handler(srv))
	r.GET("/api/admin/users", _UserService_ListUsers0_HTTP_Handler(srv))
	r.POST("/api/admin/users/{user_id}/role", _UserService_AdminUpdateUserRole0_HTTP_Handler(srv))
	r.POST("/api/admin/users/{user_id}/status", _UserService_AdminUpdateUserStatus0_HTTP_Handler(srv))
	r.POST("/api/admin/users/{user_id}/reset-password", _UserService_AdminResetUserPassword0_HTTP_Handler(srv))
}

func _UserService_Register0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RegisterRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceRegister)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Register(ctx, req.(*RegisterRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RegisterReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_Login0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceLogin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Login(ctx, req.(*LoginRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_ResetPassword0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ResetPasswordRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceResetPassword)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ResetPassword(ctx, req.(*ResetPasswordRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ResetPasswordReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_RegisterMerchant0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RegisterMerchantRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceRegisterMerchant)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RegisterMerchant(ctx, req.(*RegisterMerchantRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RegisterMerchantReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_RefreshToken0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RefreshTokenRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceRefreshToken)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RefreshToken(ctx, req.(*RefreshTokenRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RefreshTokenReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_Logout0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LogoutRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceLogout)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Logout(ctx, req.(*LogoutRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LogoutReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_GetMe0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceGetMe)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMe(ctx, req.(*GetMeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMeReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_AdminCreateUser0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AdminCreateUserRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceAdminCreateUser)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AdminCreateUser(ctx, req.(*AdminCreateUserRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AdminCreateUserReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_ListUsers0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListUsersRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceListUsers)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListUsers(ctx, req.(*ListUsersRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListUsersReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_AdminUpdateUserRole0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AdminUpdateUserRoleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceAdminUpdateUserRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AdminUpdateUserRole(ctx, req.(*AdminUpdateUserRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AdminUpdateUserRoleReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_AdminUpdateUserStatus0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AdminUpdateUserStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceAdminUpdateUserStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AdminUpdateUserStatus(ctx, req.(*AdminUpdateUserStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AdminUpdateUserStatusReply)
		return ctx.Result(200, reply)
	}
}

func _UserService_AdminResetUserPassword0_HTTP_Handler(srv UserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AdminResetUserPasswordRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserServiceAdminResetUserPassword)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AdminResetUserPassword(ctx, req.(*AdminResetUserPasswordRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AdminResetUserPasswordReply)
		return ctx.Result(200, reply)
	}
}

type UserServiceHTTPClient interface {
	// AdminCreateUser 主账号创建团队子账号；不创建买家或新的主账号。
	AdminCreateUser(ctx context.Context, req *AdminCreateUserRequest, opts ...http.CallOption) (rsp *AdminCreateUserReply, err error)
	// AdminResetUserPassword 主账号重置团队子账号密码；不允许重置买家、主账号或其他商家空间账号。
	AdminResetUserPassword(ctx context.Context, req *AdminResetUserPasswordRequest, opts ...http.CallOption) (rsp *AdminResetUserPasswordReply, err error)
	// AdminUpdateUserRole 主账号修改团队子账号角色；不允许改为买家或新的主账号。
	AdminUpdateUserRole(ctx context.Context, req *AdminUpdateUserRoleRequest, opts ...http.CallOption) (rsp *AdminUpdateUserRoleReply, err error)
	// AdminUpdateUserStatus 主账号启用或禁用团队子账号。
	AdminUpdateUserStatus(ctx context.Context, req *AdminUpdateUserStatusRequest, opts ...http.CallOption) (rsp *AdminUpdateUserStatusReply, err error)
	// GetMe 查看当前登录用户。
	GetMe(ctx context.Context, req *GetMeRequest, opts ...http.CallOption) (rsp *GetMeReply, err error)
	// ListUsers 主账号查询团队子账号列表。
	ListUsers(ctx context.Context, req *ListUsersRequest, opts ...http.CallOption) (rsp *ListUsersReply, err error)
	// Login 用户名密码登录。
	Login(ctx context.Context, req *LoginRequest, opts ...http.CallOption) (rsp *LoginReply, err error)
	// Logout 登出当前 refresh session。
	Logout(ctx context.Context, req *LogoutRequest, opts ...http.CallOption) (rsp *LogoutReply, err error)
	// RefreshToken 刷新 access token，并轮换 refresh token。
	RefreshToken(ctx context.Context, req *RefreshTokenRequest, opts ...http.CallOption) (rsp *RefreshTokenReply, err error)
	// Register 公开注册买家账号。
	Register(ctx context.Context, req *RegisterRequest, opts ...http.CallOption) (rsp *RegisterReply, err error)
	// RegisterMerchant 公开注册商家/主播主账号。
	RegisterMerchant(ctx context.Context, req *RegisterMerchantRequest, opts ...http.CallOption) (rsp *RegisterMerchantReply, err error)
	// ResetPassword 原型阶段：仅凭账号重置密码，方便测试和压测准备账号。
	ResetPassword(ctx context.Context, req *ResetPasswordRequest, opts ...http.CallOption) (rsp *ResetPasswordReply, err error)
}

type UserServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewUserServiceHTTPClient(client *http.Client) UserServiceHTTPClient {
	return &UserServiceHTTPClientImpl{client}
}

// AdminCreateUser 主账号创建团队子账号；不创建买家或新的主账号。
func (c *UserServiceHTTPClientImpl) AdminCreateUser(ctx context.Context, in *AdminCreateUserRequest, opts ...http.CallOption) (*AdminCreateUserReply, error) {
	var out AdminCreateUserReply
	pattern := "/api/admin/users"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserServiceAdminCreateUser))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// AdminResetUserPassword 主账号重置团队子账号密码；不允许重置买家、主账号或其他商家空间账号。
func (c *UserServiceHTTPClientImpl) AdminResetUserPassword(ctx context.Context, in *AdminResetUserPasswordRequest, opts ...http.CallOption) (*AdminResetUserPasswordReply, error) {
	var out AdminResetUserPasswordReply
	pattern := "/api/admin/users/{user_id}/reset-password"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserServiceAdminResetUserPassword))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// AdminUpdateUserRole 主账号修改团队子账号角色；不允许改为买家或新的主账号。
func (c *UserServiceHTTPClientImpl) AdminUpdateUserRole(ctx context.Context, in *AdminUpdateUserRoleRequest, opts ...http.CallOption) (*AdminUpdateUserRoleReply, error) {
	var out AdminUpdateUserRoleReply
	pattern := "/api/admin/users/{user_id}/role"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserServiceAdminUpdateUserRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// AdminUpdateUserStatus 主账号启用或禁用团队子账号。
func (c *UserServiceHTTPClientImpl) AdminUpdateUserStatus(ctx context.Context, in *AdminUpdateUserStatusRequest, opts ...http.CallOption) (*AdminUpdateUserStatusReply, error) {
	var out AdminUpdateUserStatusReply
	pattern := "/api/admin/users/{user_id}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserServiceAdminUpdateUserStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetMe 查看当前登录用户。
func (c *UserServiceHTTPClientImpl) GetMe(ctx context.Context, in *GetMeRequest, opts ...http.CallOption) (*GetMeReply, error) {
	var out GetMeReply
	pattern := "/api/users/me"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserServiceGetMe))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListUsers 主账号查询团队子账号列表。
func (c *UserServiceHTTPClientImpl) ListUsers(ctx context.Context, in *ListUsersRequest, opts ...http.CallOption) (*ListUsersReply, error) {
	var out ListUsersReply
	pattern := "/api/admin/users"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserServiceListUsers))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Login 用户名密码登录。
func (c *UserServiceHTTPClientImpl) Login(ctx context.Context, in *LoginRequest, opts ...http.CallOption) (*LoginReply, error) {
	var out LoginReply
	pattern := "/api/users/login"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserServiceLogin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Logout 登出当前 refresh session。
func (c *UserServiceHTTPClientImpl) Logout(ctx context.Context, in *LogoutRequest, opts ...http.CallOption) (*LogoutReply, error) {
	var out LogoutReply
	pattern := "/api/users/logout"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserServiceLogout))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// RefreshToken 刷新 access token，并轮换 refresh token。
func (c *UserServiceHTTPClientImpl) RefreshToken(ctx context.Context, in *RefreshTokenRequest, opts ...http.CallOption) (*RefreshTokenReply, error) {
	var out RefreshTokenReply
	pattern := "/api/users/refresh"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserServiceRefreshToken))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Register 公开注册买家账号。
func (c *UserServiceHTTPClientImpl) Register(ctx context.Context, in *RegisterRequest, opts ...http.CallOption) (*RegisterReply, error) {
	var out RegisterReply
	pattern := "/api/users/register"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserServiceRegister))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// RegisterMerchant 公开注册商家/主播主账号。
func (c *UserServiceHTTPClientImpl) RegisterMerchant(ctx context.Context, in *RegisterMerchantRequest, opts ...http.CallOption) (*RegisterMerchantReply, error) {
	var out RegisterMerchantReply
	pattern := "/api/merchants/register"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserServiceRegisterMerchant))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ResetPassword 原型阶段：仅凭账号重置密码，方便测试和压测准备账号。
func (c *UserServiceHTTPClientImpl) ResetPassword(ctx context.Context, in *ResetPasswordRequest, opts ...http.CallOption) (*ResetPasswordReply, error) {
	var out ResetPasswordReply
	pattern := "/api/users/reset-password"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserServiceResetPassword))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}