# live-auction-bid-backend/openapi/auction.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Live Auction Proto HTTP API",
    "version": "2026-06-16",
    "description": "OpenAPI contract synchronized with api/auction/service/v1/*.proto HTTP bindings. Business JSON routes are implemented by Kratos protoc-gen-go-http handlers; upload, websocket, health, metrics, worker and cluster endpoints remain hand-written infrastructure routes."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/users/register": {
      "post": {
        "operationId": "register",
        "summary": "Register buyer user",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthReply"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "operationId": "login",
        "summary": "Login user",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthReply"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "summary": "Reset password",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserReply"
                }
              }
            }
          }
        }
      }
    },
    "/api/merchants/register": {
      "post": {
        "operationId": "registerMerchant",
        "summary": "Register merchant account",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterMerchantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthReply"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/refresh": {
      "post": {
        "operationId": "refreshToken",
        "summary": "Refresh access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "result"
                  ],
                  "properties": {
                    "result": {
                      "$ref": "#/components/schemas/ReplyResult"
                    },
                    "tokens": {
                      "$ref": "#/components/schemas/AuthTokens"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Logout user session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyReply"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "operationId": "listAdminUsers",
        "summary": "List team users for current main account",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "roleCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserStatus"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersReply"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "adminCreateUser",
        "summary": "Create a scoped team user for current main account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserReply"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}/role": {
      "post": {
        "operationId": "adminUpdateUserRole",
        "summary": "Update a scoped team user role",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserReply"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}/status": {
      "post": {
        "operationId": "adminUpdateUserStatus",
        "summary": "Update a scoped team user status",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserReply"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}/reset-password": {
      "post": {
        "operationId": "adminResetUserPassword",
        "summary": "Reset a scoped team user password",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminResetUserPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unified result envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserReply"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ReplyResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          }
        }
      },
...
      "UserStatus": {
        "type": "string",
        "enum": [
          "USER_STATUS_UNSPECIFIED",
          "USER_STATUS_ACTIVE",
          "USER_STATUS_DISABLED"
        ]
      },
...
      "AuthTokens": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accessToken",
          "refreshToken"
        ],
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "accessExpiresAtUnixMs": {
            "$ref": "#/components/schemas/Int64"
          },
          "refreshExpiresAtUnixMs": {
            "$ref": "#/components/schemas/Int64"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "username",
          "password",
          "nickname"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          }
        }
      },
      "RegisterMerchantRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "merchantName": {
            "type": "string"
          }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "LogoutRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "AuthReply": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "result"
        ],
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ReplyResult"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "tokens": {
            "$ref": "#/components/schemas/AuthTokens"
          }
        }
      },
      "UserReply": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "result"
        ],
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ReplyResult"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "EmptyReply": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "result"
        ],
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ReplyResult"
          }
        }
      },
...
    }
  }
}